backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.0
